fix(outdir): fix bug caused by bad outdir management when in dry run mode

+++ src/utils/IO/WriteFileRepo.ts
@@ -1,9 +1,15 @@
 import * as fs from "fs";
+import { CreateTemplateConfig, createTemplateFolderConfig } from "../../config";
 import { encoding } from "../../consts";
 import { Executor } from "../../types";
 
 export class WriteFileRepo implements Executor {
+  constructor(
+    private config: CreateTemplateConfig = createTemplateFolderConfig
+  ) {}
   async execute(realPath: string, content: string): Promise<void> {
-    return fs.writeFileSync(realPath, content, encoding);
+    if (!this.config.dryRun) {
+      return fs.writeFileSync(realPath, content, encoding);
+    }
   }
 }
